requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pyh9f0ad1d_0
audioread=3.0.0=py310h5588dad_1
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
brotlipy=0.7.0=py310h8d17308_1005
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.9.24=h5b45459_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310h628cb3f_2
charset-normalizer=2.1.1=pyhd8ed1ab_0
contourpy=1.0.6=py310h232114e_0
cryptography=38.0.3=py310h6e82f81_0
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
fonttools=4.38.0=py310h8d17308_1
freetype=2.12.1=h546665d_0
idna=3.4=pyhd8ed1ab_0
intel-openmp=2022.1.0=h57928b3_3787
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
kiwisolver=1.4.4=py310h232114e_1
lame=3.100=hcfcfb64_1003
lcms2=2.14=h90d422f_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=16_win64_mkl
libdeflate=1.14=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libflac=1.4.2=h63175ca_0
liblapack=3.9.0=16_win64_mkl
libogg=1.3.4=h8ffe710_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.38=h19919ed_0
librosa=0.9.2=pyhd8ed1ab_0
libsndfile=1.1.0=h63175ca_0
libsqlite=3.39.4=hcfcfb64_0
libtiff=4.4.0=h8e97e67_4
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libzlib=1.2.13=hcfcfb64_4
llvmlite=0.39.1=py310hb84602e_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib-base=3.6.2=py310h51140c5_0
mkl=2022.1.0=h6a75c08_874
mpg123=1.30.2=h63175ca_1
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
numba=0.56.3=py310h19bcfe9_0
numpy=1.23.4=py310h4a8f9c9_1
openjpeg=2.5.0=hc9384bd_1
openssl=3.0.7=hcfcfb64_0
packaging=21.3=pyhd8ed1ab_0
pillow=9.2.0=py310hd4fb230_3
pip=22.3.1=pyhd8ed1ab_0
pooch=1.6.0=pyhd8ed1ab_0
pthread-stubs=0.4=hcd874cb_1001
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
pysoundfile=0.11.0=pyhd8ed1ab_0
python=3.10.6=hcf16a7b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
requests=2.28.1=pyhd8ed1ab_1
resampy=0.4.2=pyhd8ed1ab_0
scikit-learn=1.1.3=py310had3394f_1
scipy=1.9.3=py310h578b7cb_2
setuptools=65.5.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
tbb=2021.6.0=h91493d7_1
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
tzdata=2022f=h191b570_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.0.0=py310h8d17308_0
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.3=h3d8a991_9
vs2015_runtime=14.32.31332=h1d6e394_9
wheel=0.38.4=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
zstd=1.5.2=h7755175_4
